+++ main/src/main/java/net/oneandone/stool/setup/Home.java
@@ -21,7 +21,7 @@ import net.oneandone.setenv.Setenv;
 import net.oneandone.stool.cli.Main;
 import net.oneandone.stool.configuration.Autoconf;
 import net.oneandone.stool.configuration.StoolConfiguration;
-import net.oneandone.stool.extensions.ExtensionsFactory;
+import net.oneandone.stool.templates.ExtensionsFactory;
 import net.oneandone.stool.util.Environment;
 import net.oneandone.stool.util.RmRfThread;
 import net.oneandone.stool.util.Session;