+++ main/src/main/java/net/oneandone/stool/stage/Stage.java
@@ -25,7 +25,7 @@ import net.oneandone.stool.cli.Main;
 import net.oneandone.stool.configuration.StageConfiguration;
 import net.oneandone.stool.docker.BuildError;
 import net.oneandone.stool.docker.Engine;
-import net.oneandone.stool.extensions.Extensions;
+import net.oneandone.stool.templates.Extensions;
 import net.oneandone.stool.scm.Scm;
 import net.oneandone.stool.ssl.KeyStore;
 import net.oneandone.stool.stage.artifact.Changes;